% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\docType{class}
\name{clustering-class}
\alias{clustering-class}
\alias{clustering}
\title{clustering}
\description{
An S4 class storing information related to a clustering of scRNAseq data.
}
\section{Slots}{

\describe{
\item{\code{Seurat_object}}{The Seurat object output of a clustering pipeline run
in Seurat. This output--commonly saved locally using the Seurat::saveRDS
command--should bind the data produced by running all steps in a standard
clustering pipeline up to and including non-linear dimensionality reduction
(see "https://satijalab.org/seurat/articles/pbmc3k_tutorial.html").}

\item{\code{clustering_resolution}}{The resolution parameter used in that pipeline.
This must be some real number between 0 and 10 (inclusive).}

\item{\code{dims}}{The dimensions of reduction used in that pipeline. This must be
some integer between 1 and 100 (inclusive).}
}}

