% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusterings.R
\name{compare_clusterings}
\alias{compare_clusterings}
\title{compare_clusterings}
\usage{
compare_clusterings(first_clustering, second_clustering)
}
\arguments{
\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}

\item{second_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #2 clustering.}
}
\value{
An HTML summary report comparing \code{first_clustering} and
\code{second_clustering}. See the README for this package for more details
regarding the figures generated here.
}
\description{
Generates an HTML summary report comparing the user's Reference Genome #1
clustering to their Reference Genome #2 clustering.
}
