% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_biadjacency_matrix.R
\name{compute_biadjacency_matrix}
\alias{compute_biadjacency_matrix}
\title{compute_biadjacency_matrix}
\usage{
compute_biadjacency_matrix(first_clustering, second_clustering)
}
\arguments{
\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}

\item{second_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #2 clustering.}
}
\value{
A biadjacency matrix between the clusters in \code{first_clustering}
and those in \code{second_clustering}.
}
\description{
Computes a biadjacency matrix between the clusters in the user's Reference
Genome #1 clustering and those in their Reference Genome #2 clustering. Each
entry i,j in that matrix is the Jaccard similarity score between cluster i in
the user's Reference Genome #1 clustering and cluster j in their Reference
Genome #2 clustering. Jaccard similarity is computed on the basis of cell IDs
shared between clusters. Cluster labels are replaced to be Seurat-compliant.
}
