% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity.R
\name{compute_similarity}
\alias{compute_similarity}
\title{compute_similarity}
\usage{
compute_similarity(first_clustering, second_clustering)
}
\arguments{
\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}

\item{second_clustering}{The S4 \code{clustering} class storing information
related to a given Reference Genome #2 clustering.}
}
\value{
The adjusted Rand index (ARI) similarity score between
\code{first_clustering} and \code{second_clustering}.
}
\description{
Computes the similarity between two different clusterings of the same
reads--one using Reference Genome #1 and the other using Reference Genome
#2--as an adjusted Rand index (ARI) similarity score.
}
