% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matching_clusters.R
\name{find_matching_clusters}
\alias{find_matching_clusters}
\title{find_matching_clusters}
\usage{
find_matching_clusters(first_clustering, second_clustering)
}
\arguments{
\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}

\item{second_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #2 clustering.}
}
\value{
A vector with a length equal to the number of clusters in
\code{first_clustering}. Each element in this vector is the integer label
for the cluster in \code{second_clustering} which is most concordant with
the cluster labelled i - 1 in \code{first_clustering}, where i is the index
of the given vector element.
}
\description{
Finds concordant ("matching") clusters betwen the user's Reference Genome #1
clustering and their Reference Genome #2 clustering by maximum weight
bipartite graph matching. See the README for this package for more details
regarding the mathematics of this algorithm.
}
