% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_cells_by_cluster}
\alias{get_cells_by_cluster}
\alias{get_cells_by_cluster,clustering-method}
\title{get_cells_by_cluster}
\usage{
get_cells_by_cluster(object)

\S4method{get_cells_by_cluster}{clustering}(object)
}
\arguments{
\item{object}{An S4 \code{clustering} class.}
}
\value{
A list of lists; each sublist is a list of cell IDs belonging to the
cluster numbered i - 1 in this \code{clustering}, where i is the sublist's
index in the list.
}
\description{
Returns a list of cells belonging to each cluster in this \code{clustering}.
}
