% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class_labels.R
\name{get_class_labels}
\alias{get_class_labels}
\title{get_class_labels}
\usage{
get_class_labels(first_clustering, second_clustering)
}
\arguments{
\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}

\item{second_clustering}{The S4 \code{clustering} class storing information
related to a given Reference Genome #2 clustering.}
}
\value{
A list of length 2; the first element is a vector of integer class
labels for \code{first_clustering}; the second a vector of integer class
labels for \code{second_clustering}.
}
\description{
Assigns class labels to the set of cells common to the user's Reference
Genome #1 clustering and the given Reference Genome #2 clustering; denoting
every cell in that set as belonging to cluster i in the user's Reference
Genome #1 clustering, and cluster j in the given Reference Genome #2
clustering.
}
