% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_clustering_resolution}
\alias{get_clustering_resolution}
\alias{get_clustering_resolution,clustering-method}
\title{get_clustering_resolution}
\usage{
get_clustering_resolution(object)

\S4method{get_clustering_resolution}{clustering}(object)
}
\arguments{
\item{object}{An S4 \code{clustering} class.}
}
\value{
The \code{clustering_resolution} slot stored by this class.
}
\description{
Gets the clustering resolution used to resolve this \code{clustering}.
}
