% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_dims}
\alias{get_dims}
\alias{get_dims,clustering-method}
\title{get_dims}
\usage{
get_dims(object)

\S4method{get_dims}{clustering}(object)
}
\arguments{
\item{object}{An S4 \code{clustering} class.}
}
\value{
The \code{dims} slot stored by this class.
}
\description{
Gets the dimensions of reduction used in this \code{clustering}.
}
