% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_clustering.R
\name{get_first_clustering}
\alias{get_first_clustering}
\title{get_first_clustering}
\usage{
get_first_clustering(first_directory, reference_resolution, dims)
}
\arguments{
\item{first_directory}{The file path to the user's local directory
containing the (1) barcodes/cell IDs, (2) features/gene IDs, and (3) count
matrix outputted by the alignment of reads to Reference Genome #1.}

\item{reference_resolution}{The resolution at which to cluster the reads.}

\item{dims}{(optional) The dimensions of reduction with which to cluster the
reads. If not selected by the user, this parameter is set to 30 by default.}
}
\value{
An S4 \code{clustering} class storing information related to the
user's Reference Genome #1 clustering.
}
\description{
Generates an S4 \code{clustering} class storing information related to
the clustering of reads aligned to the user's Reference Genome #1.
}
