% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_similar_clustering.R
\name{get_max_similar_clustering}
\alias{get_max_similar_clustering}
\title{get_max_similar_clustering}
\usage{
get_max_similar_clustering(
  first_clustering,
  nearest_neighbors,
  dims,
  init_resolution,
  increment
)
}
\arguments{
\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}

\item{nearest_neighbors}{The nearest neighbors graph construction computed
from the alignment of reads to Reference Genome #2.}

\item{dims}{The dimensions of reduction with which to cluster those reads.}

\item{init_resolution}{The initial resolution at which to cluster those
reads.}

\item{increment}{The increment at which to cluster iteratively at increasing
resolutions beginning with \code{init_resolution}.}
}
\value{
An S4 \code{clustering} class storing information related to that
Reference Genome #2 clustering which is maximally similar to the user's
\code{first_clustering} with an error of +/- \code{increment} resolutions.
}
\description{
Clusters the reads aligned to Reference Genome #2 iteratively at increasing
increments of \code{increment} resolutions; starting at
\code{init_resolution} and stopping only when the adjusted Rand index (ARI)
similarity measure between the user's Reference Genome #1 clustering and the
given Reference Genome #2 clustering has been globally maximized with an
error of +/- \code{increment} resolutions. See the README for this package
for more details regarding the mathematics of this algorithm.
}
