% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_number_of_clusters}
\alias{get_number_of_clusters}
\alias{get_number_of_clusters,clustering-method}
\title{get_number_of_clusters}
\usage{
get_number_of_clusters(object)

\S4method{get_number_of_clusters}{clustering}(object)
}
\arguments{
\item{object}{An S4 \code{clustering} class.}
}
\value{
The number of clusters in this \code{clustering}.
}
\description{
Gets the number of clusters in this \code{clustering}.
}
