% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_second_clustering.R
\name{get_second_clustering}
\alias{get_second_clustering}
\title{get_second_clustering}
\usage{
get_second_clustering(second_directory, first_clustering)
}
\arguments{
\item{second_directory}{The file path to the user's local directory
containing the (1) barcodes/cell IDs, (2) features/gene IDs, and (3) count
matrix outputted by the alignment of reads to Reference Genome #2.}

\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}
}
\value{
An S4 \code{clustering} class storing information related to the
user's Reference Genome #2 clustering.
}
\description{
Generates an S4 \code{clustering} class storing information related to
the clustering of reads aligned to the user's Reference Genome #2. The
resolution of this clustering is optimized within an error of +/- 0.01 so as
to maximize its similarity to the user's Reference Genome #1 clustering.
}
