% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_Seurat_object}
\alias{get_Seurat_object}
\alias{get_Seurat_object,clustering-method}
\title{get_Seurat_object}
\usage{
get_Seurat_object(object)

\S4method{get_Seurat_object}{clustering}(object)
}
\arguments{
\item{object}{An S4 \code{clustering} class.}
}
\value{
The \code{Seurat_object} slot stored by this class.
}
\description{
Gets the Seurat object associated with this \code{clustering}.
}
