% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unique_cells.R
\name{remove_unique_cells}
\alias{remove_unique_cells}
\title{remove_unique_cells}
\usage{
remove_unique_cells(first_clustering, second_clustering)
}
\arguments{
\item{first_clustering}{The S4 \code{clustering} class storing information
related to the user's Reference Genome #1 clustering.}

\item{second_clustering}{The S4 \code{clustering} class storing information
related to a given Reference Genome #2 clustering.}
}
\value{
A list of length 2; the first element is a list of non-unique cell
IDs by cluster in \code{first_clustering}; the second a list of a non-unique
cell IDs by cluster in \code{second_clustering}.
}
\description{
Removes any cells from two given clusterings which are not shared between
those clusterings. This allows a similarity score between these clusterings
to be computed over a set of cells common to both clusterings.
}
